updater: Trigger heroku build even if bot is up-to-date
`.update deploy` works by force-pushing the latest upstream state to
Heroku, which then triggers a new build. This updates both the docker
container and the userbot code. If the userbot itself is up-to-date, we
can't update the docker container, as force-pushing will just result in
"Everything is up-to-date".

Amending a commit without modifying the message or contents will still
result in a different commit ID, which can then be force-pushed to
Heroku to trigger a build.

Co-authored-by: KenHV <yo@kenharris.xyz>
Co-authored-by: mrismanaziz <mrismanaziz@gmail.com>